move car_tests into car_cmd
